requirements.txt
# ==============================================
# Mocap Pro - Production Requirements
# Версия: 1.0.0
# Автор: Mocap Pro Team
# ==============================================

# ========== ОСНОВНЫЕ ЗАВИСИМОСТИ ==========

# Компьютерное зрение и обработка видео
opencv-python>=4.8.1,<5.0.0
opencv-contrib-python>=4.8.1,<5.0.0
mediapipe>=0.10.0,<0.11.0

# Графический интерфейс
PyQt6>=6.6.0,<7.0.0
PyQt6-Qt6>=6.6.0,<7.0.0
PyQt6-sip>=13.6.0,<14.0.0
PyOpenGL>=3.1.6,<4.0.0
PyOpenGL-accelerate>=3.1.6,<4.0.0

# Научные вычисления и математика
numpy>=1.24.0,<2.0.0
scipy>=1.11.0,<2.0.0
pandas>=2.0.0,<3.0.0
transforms3d>=0.4.1,<0.5.0

# Машинное обучение и фильтрация
scikit-learn>=1.3.0,<2.0.0
filterpy>=1.4.5,<2.0.0

# ========== РАБОТА С ДАННЫМИ ==========

# Форматы файлов
PyYAML>=6.0,<7.0
python-rapidjson>=1.10,<2.0
openpyxl>=3.1.0,<4.0.0  # Для Excel экспорта

# Сериализация
msgpack>=1.0.5,<2.0.0
protobuf>=4.24.0,<5.0.0  # Для MediaPipe

# ========== ГРАФИКА И ВИЗУАЛИЗАЦИЯ ==========

# 3D графика
pyopengl>=3.1.6,<4.0.0
vispy>=0.13.0,<0.14.0  # Для продвинутой визуализации

# Графики и диаграммы
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<0.13.0
plotly>=5.17.0,<6.0.0

# Обработка изображений
Pillow>=10.0.0,<11.0.0
imageio>=2.31.0,<3.0.0
imageio-ffmpeg>=0.4.8,<0.5.0

# ========== ПРОИЗВОДИТЕЛЬНОСТЬ ==========

# Ускорение вычислений
numba>=0.58.0,<0.59.0
llvmlite>=0.41.0,<0.42.0

# Параллельные вычисления
joblib>=1.3.0,<2.0.0
threadpoolctl>=3.2.0,<4.0.0

# Кэширование
cachetools>=5.3.0,<6.0.0
diskcache>=5.6.0,<6.0.0

# ========== СЕТЬ И КОММУНИКАЦИЯ ==========

# Сетевое взаимодействие
websockets>=12.0,<13.0
aiohttp>=3.9.0,<4.0.0
requests>=2.31.0,<3.0.0

# Сериализация для сети
msgpack>=1.0.5,<2.0.0

# ========== ИНТЕГРАЦИЯ С 3D ПАКЕТАМИ ==========

# Blender интеграция
bpy>=3.6.0,<4.0.0; platform_system != "Windows"  # Только для Linux/macOS

# Unity/Unreal (через промежуточные форматы)
pytorch3d>=0.7.5,<0.8.0  # Для 3D операций
trimesh>=3.23.0,<4.0.0   # Для работы с мешами

# ========== АУДИО И ВИДЕО ==========

# Обработка видео
moviepy>=1.0.3,<2.0.0
ffmpeg-python>=0.2.0,<0.3.0

# Аудио обработка
pydub>=0.25.1,<0.26.0
librosa>=0.10.0,<0.11.0

# ========== ИНСТРУМЕНТЫ РАЗРАБОТКИ ==========

# Логирование
loguru>=0.7.0,<0.8.0
colorlog>=6.7.0,<7.0.0

# Конфигурация
pydantic>=2.4.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Тестирование
pytest>=7.4.0,<8.0.0
pytest-qt>=4.2.0,<5.0.0
pytest-cov>=4.1.0,<5.0.0

# Профилирование
pyinstrument>=4.6.0,<5.0.0
memory-profiler>=0.61.0,<0.62.0

# Типизация
typing-extensions>=4.8.0,<5.0.0
mypy>=1.6.0,<2.0.0

# ========== БЕЗОПАСНОСТЬ ==========

# Шифрование и безопасность
cryptography>=41.0.0,<42.0.0
python-jose>=3.3.0,<4.0.0

# ========== ПЛАТФОРМО-ЗАВИСИМЫЕ ==========

# Windows специфичные
pywin32>=306,<307; platform_system == "Windows"
comtypes>=1.2.0,<2.0.0; platform_system == "Windows"

# macOS специфичные
pyobjc>=9.2,<10.0; platform_system == "Darwin"

# Linux специфичные
pycairo>=1.25.0,<2.0.0; platform_system == "Linux"
dbus-python>=1.3.2,<2.0.0; platform_system == "Linux"

# ========== ДОПОЛНИТЕЛЬНЫЕ УТИЛИТЫ ==========

# Дата и время
python-dateutil>=2.8.0,<3.0.0
pytz>=2023.3,<2024.0

# Работа с путями
pathlib2>=2.3.0,<3.0.0; python_version < "3.10"

# Уведомления
plyer>=2.1.0,<3.0.0
desktop-notifier>=3.5.0,<4.0.0

# Интернационализация
babel>=2.12.0,<3.0.0

# ========== ОПТИМИЗАЦИЯ ПАМЯТИ ==========

# Экономное использование памяти
zarr>=2.16.0,<3.0.0
h5py>=3.9.0,<4.0.0

# ========== РАБОЧИЕ ВЕРСИИ (ПРОТЕСТИРОВАННЫЕ) ==========

# Эти версии точно работают вместе
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
mediapipe==0.10.8
PyQt6==6.6.0
numpy==1.24.3
scipy==1.11.3
pandas==2.0.3
transforms3d==0.4.1
scikit-learn==1.3.0
filterpy==1.4.5
PyYAML==6.0.1
python-rapidjson==1.10
msgpack==1.0.5
protobuf==4.24.4
pyopengl==3.1.7
matplotlib==3.7.2
Pillow==10.0.0
numba==0.58.0
cachetools==5.3.2
requests==2.31.0
loguru==0.7.2
pydantic==2.4.2
python-dateutil==2.8.2

# ==============================================
# ИНСТРУКЦИЯ ПО УСТАНОВКЕ:
# ==============================================

# БАЗОВАЯ УСТАНОВКА (минимальные требования):
# pip install -r requirements.txt

# ПОЛНАЯ УСТАНОВКА (все зависимости):
# pip install -r requirements.txt --upgrade

# УСТАНОВКА ДЛЯ РАЗРАБОТЧИКОВ:
# pip install -r requirements-dev.txt  # если есть отдельный файл

# ==============================================
# ПРИМЕЧАНИЯ:
# ==============================================

# 1. OpenCV с contrib модулями требуется для ARUCO и дополнительных алгоритмов
# 2. MediaPipe 0.10.x стабильная версия для трекинга позы
# 3. PyQt6 6.6.0+ для поддержки всех функций интерфейса
# 4. NumPy 1.24+ для совместимости с новыми версиями SciPy
# 5. Transform3d для 3D преобразований и кватернионов

# ==============================================
# АЛЬТЕРНАТИВНЫЕ ВЕРСИИ (если проблемы):
# ==============================================

# Для старых систем:
# opencv-python==4.5.5.64
# mediapipe==0.8.11
# PyQt5==5.15.9
# numpy==1.21.6

# Для Windows с Python 3.11:
# pip install "opencv-python<4.9" "mediapipe<0.10.9"

# ==============================================
# ВОЗМОЖНЫЕ КОНФЛИКТЫ:
# ==============================================

# 1. NumPy и SciPy должны быть совместимыми версиями
# 2. Protobuf может конфликтовать с другими пакетами
# 3. PyQt6 не совместим с PyQt5 в одном окружении

# ==============================================
# ПРОВЕРКА УСТАНОВКИ:
# ==============================================

# После установки проверьте:
# python -c "import cv2; print(f'OpenCV: {cv2.__version__}')"
# python -c "import mediapipe as mp; print(f'MediaPipe: {mp.__version__}')"
# python -c "from PyQt6.QtCore import QT_VERSION_STR; print(f'Qt: {QT_VERSION_STR}')"